changed the signature of the basis functions to make k_x and k_y UInt64

+++ homework_5/math_functions.jl
@@ -4,7 +4,7 @@ include("constants.jl")
 returns the value of the basis function `f_k_x`
     for the given x-coordinate `x` and wave number `k_x`
 """
-function f_k_x(x::Float64, k_x::Float64)::Float64
+function f_k_x(x::Float64, k_x::UInt64)::Float64
     return sqrt(2 / Lx) * sin(k_x * π * x / Lx)
 end
 
@@ -12,6 +12,9 @@ end
 returns the value of the basis function `g_k_y`
     for the given y-coordinate `y` and wave number `k_y`
 """
-function g_k_y(y::Float64, k_y::Float64)::Float64
+function g_k_y(y::Float64, k_y::UInt64)::Float64
     return sqrt(2 / Ly) * sin(k_y * π * y / Ly)
 end
+
+
+
